feat: Improve performance by adding Post Slice and modifying code in HomePg, AllPostsPg, and addPostPg
This commit adds the Post Slice feature to improve performance. It also includes code modifications in the HomePg, AllPostsPg, and addPostPg components to meet the requirements of the feature. Additionally, it changes the header navbar item from "All posts" to "Your Posts".

Refactor the `AuthLayout` component to update the authentication logic. Instead of redirecting to the home page when `authStatus` is true and `authentication` is true, it now redirects to the login page. Similarly, when `authStatus` is false and `authentication` is false, it redirects to the home page. This change ensures that the appropriate page is displayed based on the authentication status.

Also, update the usage of the `Protected` component in the `main.jsx` file to pass the `authentication` prop explicitly.

This commit addresses the issue of incorrect redirection in the `AuthLayout` component and improves the authentication flow in the application.

chore: Fix import paths and update npm dependencies

Working

Merge branch 'main' of https://github.com/mhamsha/Blog-App

Add File to git ignore

Merge pull request #3 from mhamsha/PagesWork

Add Pages

Merge pull request #2 from mhamsha/components-work

Components work

add postCardComp