feat: Remove unnecessary import from appwriteConfig.js and ContainerComp.jsx
Remove the unused import of the `Query` object from the `appwriteConfig.js` file. This import is no longer needed after the recent code changes.

In the `ContainerComp.jsx` file, remove the import of the `React` library. This import is unnecessary as the component does not use any React-specific features.

These changes help to clean up the code and remove unnecessary dependencies.