refactor: Update PostPg component to optimize data fetching
The code changes in the PostPg component optimize the data fetching process by using a more efficient approach. Instead of fetching all posts, the code now checks if the searchPostsStore or yourPostsStore is available and fetches the corresponding posts. This change improves the performance and efficiency of the data fetching process in the PostPg component.